Patrick need food badly
Named in honour of the first person to starve to death. Added
rudimentary hunger mechanic, as well as an actual advantage to the
survival skill (which nobody actually has yet) and a bonus to hunting
for ranged weapon users. Hungry people fight worse and will eventually
die, although the hungrier you are the slower your hunger drops, so
usually only one or two people will actually die of hunger. (keeps
things interesting.)